backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.18
aiofiles==24.1.0
jinja2==3.1.4
pydantic==2.10.0
python-magic==0.4.27
werkzeug==3.1.3
pytest==8.3.4
httpx==0.28.1
